testing: fixed insertion of VJOURNAL
WebDAVSource, in contrast to CalDAVSource, does not check for UID
conflicts before the PUT. As a result, testSimpleInsert and
testLocalDeleteAll were failing when the VJOURNAL test item had
already been imported.

Need to strip the UID in cases where it is not guaranteed that the
item is not already in the database. Cases where that can be taken
for granted because of a preceeding delete continue to use the UID
embedded in the test data (but don't rely on it, because update()
without LUID will find the existing item and use its LUID).