D-Bus server: use weak pointers with asynchronous D-Bus calls
An asynchronous D-Bus call will invoke the callback even if
the call instance itself was already deleted. Therefore binding
the this pointer of the call instance owner is not safe. This
caused use-after-free errors in local sync (found during testing).

Use weak pointers as a precaution where it can be done easily.
Some remaining start() calls still need to be converted.