testing: fixed testAddBothSides for VJOURNAL
"davfile" (= WebDAV on client, file on server) failed for testAddBothSides
and testAddBothSidesRefresh because the DESCRIPTION was different and
was merged on the client, which caused an unexpected update to be sent
to the server instead of just the delete.

Solved as for other kinds of data: avoid the merging by keeping the data
identical.