testing: added googleeas = Google Contacts via ActiveSync
Testing shows that compared to Exchange, some additional problems
exist, which are ignored via synccompare:
- FN gets overwritten.
- CATGEORIES are lost.

Not sure whether anything can be done, so let's consider this
permanently broken.

BDAY also gets modified. Perhaps that can be fixed, so a synccompare
workaround will be commit separately so that it can be reverted later.