testing: ignore some known ActiveSync problems
Exchange disagrees with SyncEvolution regarding some timezones.
Ignore for now.

testReadItem404 with Google fails differently than expected
because Google does not support the Fetch operation.

testExtensions fails with both because the synccompare tweaks for
ActiveSync get disabled in that test.

One way syncing is not supported for ActiveSync (depends on old
sync keys, which tend to be no longer supported by the ActiveSync
server).