folks: initial set of classes for viewing, sorting, searching
The classes on top of libfolks are  designed such that they can
be tested without involving libfolks. Right now the goal is only
to get the API right; most of the actual functionality is missing.

The terminology (added/modified/removed) is aligned with libebook
terminology and the PIM Manager API description.