folks: added support for retrieving contacts
The choice of std::vector in getContacts() comes from the goal of
using this function more or less directly in a D-Bus implementation.
It works efficiently because the size can be computed in advance.

A more general API, like for example one which stores via some append
iterator, would also have been possible but probably isn't worth
the effort.