util: add GetWithDef()
Looks up a value in a map and returns a found value or a default. This
mimicks Python's dict.get().

Because it is not possible in C++ to return None as default, the
returned value is wrapped in a InitState to tell the caller whether
the value was found.

Depends on the unified InitState template.