local sync: allow sync and target context to be the same
It is possible to sync source A in context @X with source B in the
same context @X. The PIM manager is doing that with one context per
peer and databases that are guaranteed to not conflict with anything
else.

Therefore allow such configs. A better check is left for later, if it
really should be needed at all.