PIM testing: use real phone in testSync()
If TEST_DBUS_PBAP_PHONE is set to the Bluetooth MAC address (like
A0:4E:04:1E:AD:30) of a phone, PBAP syncing with that phone is tested.

The phone must be paired, connected and support SyncML in addition to
PBAP. The test hard-codes Nokia SyncML settings to keep it simpe and
because Nokia phones are most likely to be used.

X- extensions added by phones as part of the local->SyncML->PBAP
roundtrip are ignored. This can hide issues. For example, the Nokia
N97 Mini incorrectly sends X-CLASS:private in vCard 3.0 mode (should
use CLASS), which is faithfully copied into the local side as-is,
althought it might be better to recognize and translate it. Not a big
deal for X-CLASS, but might be for other properties.