D-Bus testing: improved test timeouts and looping in GLib
It was observed that raising an exception in the signal handler did
not propagate to the calling Python code when it was currently in
loop.get_context().iteration().

This commit adds debug logging for timeouts and some utility methods
which avoid the problem by adding logging to the iteration()
call. Python then re-raises the exception in the D-Bus logging code.

This looks like a deficiency (bug?!) of the Python bindings for
iteration() which is merely worked around.

However, the utility methods may also be useful by their own right,
for example to avoid the awkward loop.run() + loop.quit() pairs.