WebDAV: auto-discovery fix
With Google Contact + CardDAV the auto-discovery failed after
finding the default address, without reporting that result.

The reason was that the search continued at the root of the server
where PROPFIND triggers an error when using the Google server. Because
of a missing check for "have result", that error was treated as
something which needs to be reported to the user.

Fixed by unifying the various checks in a singe class.